+++ pkg/block/adapter.go
@@ -128,20 +128,11 @@ type Properties struct {
 	StorageClass *string
 }
 
-// WalkFunc is called for each object visited by the Walk.
-// The id argument contains the argument to Walk as a prefix; that is, if Walk is called with "test/data/",
-// which is a prefix containing the object "test/data/a", the walk function will be called with argument "test/data/a".
-// If there was a problem walking to the object, the incoming error will describe the problem and the function can decide
-// how to handle that error.
-// If an error is returned, processing stops.
-type WalkFunc func(id string) error
-
 type Adapter interface {
 	InventoryGenerator
 	Put(ctx context.Context, obj ObjectPointer, sizeBytes int64, reader io.Reader, opts PutOpts) error
 	Get(ctx context.Context, obj ObjectPointer, expectedSize int64) (io.ReadCloser, error)
 	GetPreSignedURL(ctx context.Context, obj ObjectPointer, mode PreSignMode) (string, error)
-	Walk(ctx context.Context, walkOpt WalkOpts, walkFn WalkFunc) error
 	Exists(ctx context.Context, obj ObjectPointer) (bool, error)
 	GetRange(ctx context.Context, obj ObjectPointer, startPosition int64, endPosition int64) (io.ReadCloser, error)
 	GetProperties(ctx context.Context, obj ObjectPointer) (Properties, error)