+++ webui/src/pages/repositories/repository/objects.jsx
@@ -1,5 +1,5 @@
-import React, {useEffect, useRef, useState} from "react";
-
+import React, {useCallback, useEffect, useRef, useState} from "react";
+import * as dayjs from 'dayjs'
 import {UploadIcon} from "@primer/octicons-react";
 import {RepositoryPageLayout} from "../../../lib/components/repository/layout";
 import RefDropdown from "../../../lib/components/repository/refDropdown";
@@ -17,10 +17,11 @@ import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
+import Alert from "react-bootstrap/Alert";
 import {BsCloudArrowUp} from "react-icons/bs";
 
 import {Tree} from "../../../lib/components/repository/tree";
-import {config, objects} from "../../../lib/api";
+import {config, objects, refs, retention, repositories, NotFoundError} from "../../../lib/api";
 import {useAPI, useAPIWithPagination} from "../../../lib/hooks/api";
 import {RefContextProvider, useRefs} from "../../../lib/hooks/repo";
 import {useRouter} from "../../../lib/hooks/router";
@@ -39,6 +40,7 @@ import { getContentType, getFileExtension, FileContents } from "./objectViewer";
 
 
 const README_FILE_NAME = 'README.md';
+const REPOSITORY_AGE_BEFORE_GC = 14;
 
 const ImportButton = ({variant = "success", enabled = false, onClick}) => {
     return (
@@ -53,6 +55,7 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
     const [numberOfImportedObjects, setNumberOfImportedObjects] = useState(0);
     const [isImportEnabled, setIsImportEnabled] = useState(false);
     const [importError, setImportError] = useState(null);
+    const [metadataFields, setMetadataFields] = useState([])
 
     const sourceRef = useRef(null);
     const destRef = useRef(null);
@@ -68,12 +71,25 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
         setImportPhase(ImportPhase.NotStarted);
         setIsImportEnabled(false);
         setNumberOfImportedObjects(0);
+        setMetadataFields([]);
     }
 
     const hide = () => {
-        if (ImportPhase.InProgress === importPhase) return;
+        if (ImportPhase.InProgress === importPhase || ImportPhase.Merging === importPhase) return;
         resetState()
         onHide()
+    };
+
+    const doMerge = async () => {
+        setImportPhase(ImportPhase.Merging);
+        try {
+            await refs.merge(repoId, importBranch, currBranch);
+            onDone();
+            hide();
+        } catch (error) {
+            setImportPhase(ImportPhase.MergeFailed);
+            setImportError(error);
+        }
     };
 
     const doImport = async () => {
@@ -83,6 +99,8 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
             setNumberOfImportedObjects(numObj);
         }
         try {
+            const metadata = {};
+            metadataFields.forEach(pair => metadata[pair.key] = pair.value)
             await runImport(
                 updateStateFromImport,
                 destRef.current.value,
@@ -90,7 +108,8 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
                 sourceRef.current.value,
                 importBranch,
                 repoId,
-                referenceId
+                referenceId,
+                metadata
             );
             onDone();
         } catch (error) {
@@ -109,7 +128,9 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
                 </Modal.Header>
                 <Modal.Body>
                     {
-                        (importPhase === ImportPhase.NotStarted || importPhase === ImportPhase.Failed) &&
+                        (importPhase === ImportPhase.NotStarted ||
+                            importPhase === ImportPhase.Failed  ||
+                            importPhase === ImportPhase.MergeFailed) &&
                         <ImportForm
                             config={config}
                             pathStyle={pathStyle}
@@ -121,6 +142,8 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
                             path={path}
                             commitMsgRef={commitMsgRef}
                             shouldAddPath={true}
+                            metadataFields={metadataFields}
+                            setMetadataFields={setMetadataFields}
                             err={importError}
                         />
                     }
@@ -138,9 +161,14 @@ const ImportModal = ({config, repoId, referenceId, referenceType, path = '', onD
                     <Button variant="secondary" disabled={importPhase === ImportPhase.InProgress} onClick={hide}>
                         Cancel
                     </Button>
-                    {
-                        <ExecuteImportButton importPhase={importPhase} importFunc={doImport} isEnabled={isImportEnabled}/>
-                    }
+
+                    <ExecuteImportButton
+                        importPhase={importPhase}
+                        importFunc={doImport}
+                        mergeFunc={doMerge}
+                        doneFunc={onDone}
+                        isEnabled={isImportEnabled}/>
+
                 </Modal.Footer>
             </Modal>
         </>
@@ -338,6 +366,49 @@ const ReadmeContainer = ({repo, reference, path='', refreshDep=''}) => {
     );
 }
 
+const NoGCRulesWarning = ({ repoId }) => {
+    const storageKey = `show_gc_warning_${repoId}`;
+    const [show, setShow] = useState(window.localStorage.getItem(storageKey) !== "false")
+    const closeAndRemember = useCallback(() => {
+        window.localStorage.setItem(storageKey, "false")
+        setShow(false)
+    }, [repoId])
+
+    const {response} = useAPI(async() => {
+        const repo = await repositories.get(repoId)
+        if (!repo.storage_namespace.startsWith('s3:') &&
+            !repo.storage_namespace.startsWith('http')) {
+            return false;
+        }
+        const createdAgo = dayjs().diff(dayjs.unix(repo.creation_date), 'days');
+        if (createdAgo > REPOSITORY_AGE_BEFORE_GC) {
+            try {
+                await retention.getGCPolicy(repoId);
+            } catch (e) {
+                if (e instanceof NotFoundError) {
+                    return true
+                }
+            }
+        }
+        return false;
+    }, [repoId]);
+
+    if (show && response) {
+        return (
+            <Alert
+                variant="warning"
+                onClose={closeAndRemember}
+                dismissible>
+                <strong>Warning</strong>: No garbage collection rules configured for this repository.
+                {' '}
+                <a href="https://docs.lakefs.io/howto/garbage-collection.html" target="_blank" rel="noreferrer">Learn More</a>.
+
+            </Alert>
+        )
+    }
+    return <></>;
+}
+
 const ObjectsBrowser = ({config, configError}) => {
     const router = useRouter();
     const {path, after} = router.query;
@@ -402,7 +473,7 @@ const ObjectsBrowser = ({config, configError}) => {
                     />
                     <ImportButton
                         onClick={() => setShowImport(true)}
-                        enabled={config.blockstore_type.import_support}
+                        enabled={config.import_support}
                     />
                     <ImportModal
                         config={config}
@@ -418,6 +489,8 @@ const ObjectsBrowser = ({config, configError}) => {
                     />
                 </ActionGroup>
             </ActionsBar>
+
+            <NoGCRulesWarning repoId={repo.id}/>
 
             <Box sx={{display: 'flex', flexDirection: 'column', gap: '10px', mb: '30px'}}>
                 <TreeContainer